+++ projects/rocprofiler-compute/src/rocprof_compute_profile/profiler_base.py
@@ -23,6 +23,7 @@ from utils.logger import (
 )
 from utils.native_tool_finder import NativeToolFinder
 from utils.utils_common import (
+    _PROFILE_OUTPUT_FORMAT,
     format_time,
     get_job_rank_and_size,
     is_only_pc_sampling,
@@ -301,6 +302,7 @@ class RocProfCompute_Base:
             # Override filter_blocks when writing profiling config yaml
             args_dict["filter_blocks"] = self._filter_blocks
             args_dict["config_dir"] = str(args_dict["config_dir"])
+            args_dict["format_rocprof_output"] = _PROFILE_OUTPUT_FORMAT
             yaml.dump(args_dict, f)
 
         # verify soc compatibility
@@ -359,7 +361,6 @@ class RocProfCompute_Base:
                 profiler_options=options,
                 workload_dir=args.output_directory,
                 loglevel=args.loglevel,
-                format_rocprof_output=args.format_rocprof_output,
                 ml_api_trace_enabled=bool(getattr(self, "_selected_frameworks", set())),
                 retain_rocpd_output=args.retain_rocpd_output,
             )