+++ projects/rocprofiler-compute/src/utils/file_io.py
@@ -19,7 +19,11 @@ from utils.logger import (
     console_warning,
     demarcate,
 )
-from utils.utils_common import canonical_config_arch, normalize_filter_to_str_list
+from utils.utils_common import (
+    _PROFILE_OUTPUT_FORMAT,
+    canonical_config_arch,
+    normalize_filter_to_str_list,
+)
 
 # TODO: use pandas chunksize or dask to read really large csv file
 # from dask import dataframe as dd
@@ -300,7 +304,7 @@ def create_df_pmc(
 
     df = pd.read_csv(pmc_perf_path)
 
-    if config_dict.get("format_rocprof_output") == "rocpd":
+    if config_dict.get("format_rocprof_output") == _PROFILE_OUTPUT_FORMAT:
         df = utils_analysis.process_rocpd_csv(df)
 
     # Demangle original KernelNames